# Reuse gRPC healthcheck client connections