This PR reuses `grpc.ClientConn` among `Check`s, it already maintains connection with backends in the background, and if it's lost it tries to reconnect based on [back-off](https://github.com/grpc/grpc/blob/master/doc/connection-backoff.md) strategy or right away when a method is invoked (like `http.Transport`). 

Having many frequent gRPC checks, especially that use TLS  may lead to serious performance penalties.

Here's a simple `Check` benchmark (insecure):

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkNewGrpcHealthProbe-12     304544        63741         -79.07%

benchmark                          old allocs     new allocs     delta
BenchmarkNewGrpcHealthProbe-12     514            177            -65.56%

benchmark                          old bytes     new bytes     delta
BenchmarkNewGrpcHealthProbe-12     233113        9649          -95.86%
```